test: split audit messages test, use nut for cmd execution
